pagination
Primary:
- Updated scripts for handling pagination of Pokemon generations 30 at a time

Secondary:
- Some class/id name-changes around the project
- Altered Pokeball animation to be a bit simpler without text, since the pagination triggers it much more often, and it became a bit too much. Load times are quicker
- Some minor style tweaks
- Made loading and errorMessage containers into Astro components. Less clutter in the HTML
- Made The <head> section into an Astro Component with parameters to pass into
