+++ src/components/ErrorMessage.astro
@@ -1,0 +1,25 @@
+---
+
+---
+
+<div id="errorMessageContainer">
+  <p id="errorMessage"></p>
+</div>
+
+<style lang="scss">
+  div {
+    display: none;
+    padding: 1em;
+    width: fit-content;
+    margin: 0 auto;
+
+    p {
+      width: fit-content;
+      border: 2px solid darkred;
+      background-color: rgb(232, 190, 190);
+      color: darkred;
+      padding: 20px;
+      font-size: 1.5em;
+    }
+  }
+</style>
