+++ src/components/Loading.astro
@@ -1,0 +1,86 @@
+---
+
+---
+
+<div class="loadingContainer">
+  <p class="loadingText">Loading Pokémons...</p>
+  <div id="loadingAnimation"></div>
+</div>
+
+<style lang="scss">
+  .loadingContainer {
+    display: none;
+    // display: flex !important;
+    flex-direction: column;
+    justify-content: center;
+    align-items: center;
+    padding: 1.5em;
+    box-shadow: 0 0 20px darkgray;
+    border-radius: 50%;
+    background-color: white;
+    position: fixed;
+    top: 50%;
+    left: 50%;
+    transform: translate(-50%, -50%);
+    z-index: 100;
+  }
+
+  .loadingText {
+    display: none;
+    font-size: 1.5rem;
+    font-weight: bold;
+  }
+
+  /* --- Loading animation --- */
+  #loadingAnimation {
+    display: flex;
+    position: relative;
+    height: 100px;
+    width: 100px;
+    background: linear-gradient(
+      to bottom,
+      rgb(254, 0, 1) 50%,
+      rgb(248, 248, 248) 50%
+    );
+    box-shadow: inset -8px -4px 0 0 rgba(0, 0, 0, 0.2);
+    border-radius: 50%;
+    border: 5px solid rgb(35, 35, 35);
+    animation: spin 1s linear infinite;
+
+    &:before {
+      content: '';
+      position: absolute;
+      height: 5px;
+      width: 100%;
+      background: rgb(35, 35, 35);
+      top: 50%;
+      transform: translatey(-50%);
+    }
+
+    &:after {
+      content: '';
+      position: absolute;
+      height: 40%;
+      width: 40%;
+      border-radius: 50%;
+      background: rgb(248, 248, 248);
+      top: 50%;
+      left: 50%;
+      transform: translate(-50%, -50%);
+      box-shadow:
+        inset 0 0 0 5px rgb(35, 35, 35),
+        inset 0 0 0 10px rgb(248, 248, 248),
+        inset 0 0 0 12px rgb(35, 35, 35);
+    }
+  }
+
+  /* Spin Animation */
+  @keyframes spin {
+    from {
+      transform: rotate(0deg);
+    }
+    to {
+      transform: rotate(360deg);
+    }
+  }
+</style>
