+++ src/css/global-styles.css
@@ -24,6 +24,13 @@ button:focus {
   box-shadow: none !important;
 }
 
+h1 {
+  text-align: center;
+  font-size: 2rem;
+  margin: 1em;
+  font-weight: 900;
+}
+
 p {
   font-size: 1rem;
   line-height: 1.6;
@@ -45,7 +52,7 @@ header {
 
 main {
   grid-area: main;
-  margin: 96px 0 0 0;
+  margin: 74px 1em 0 1em;
 }
 
 footer {
