+++ src/css/style.css
@@ -1,3 +1,4 @@
+@charset "UTF-8";
 * {
   box-sizing: border-box;
 }
@@ -24,6 +25,13 @@ button:focus {
   box-shadow: none !important;
 }
 
+h1 {
+  text-align: center;
+  font-size: 2rem;
+  margin: 1em;
+  font-weight: 900;
+}
+
 p {
   font-size: 1rem;
   line-height: 1.6;
@@ -45,7 +53,7 @@ header {
 
 main {
   grid-area: main;
-  margin: 96px 0 0 0;
+  margin: 74px 1em 0 1em;
 }
 
 footer {
@@ -85,101 +93,20 @@ html {
   color-scheme: light dark;
 }
 
-.contentContainer {
-  width: min(80em, 100% - 1rem);
-  margin-inline: auto;
+.content h1 {
+  color: var(--text-color);
 }
 
-.loadingContainer {
-  display: none;
-  flex-direction: column;
-  justify-content: center;
-  align-items: center;
-  padding: 2em;
-}
-
-.loadingText {
-  font-size: 1.5em;
-  font-weight: bold;
-}
-
-/* --- Loading animation --- */
-#loadingAnimation {
-  display: flex;
-  position: relative;
-  height: 100px;
-  width: 100px;
-  background: linear-gradient(to bottom, rgb(254, 0, 1) 50%, rgb(248, 248, 248) 50%);
-  box-shadow: inset -8px -4px 0 0 rgba(0, 0, 0, 0.2);
-  border-radius: 50%;
-  border: 5px solid rgb(35, 35, 35);
-  animation: spin 1s linear infinite;
-}
-#loadingAnimation:before {
-  content: "";
-  position: absolute;
-  height: 5px;
-  width: 100%;
-  background: rgb(35, 35, 35);
-  top: 50%;
-  transform: translatey(-50%);
-}
-#loadingAnimation:after {
-  content: "";
-  position: absolute;
-  height: 40%;
-  width: 40%;
-  border-radius: 50%;
-  background: rgb(248, 248, 248);
-  top: 50%;
-  left: 50%;
-  transform: translate(-50%, -50%);
-  box-shadow: inset 0 0 0 5px rgb(35, 35, 35), inset 0 0 0 10px rgb(248, 248, 248), inset 0 0 0 12px rgb(35, 35, 35);
-}
-
-/* Spin Animation */
-@keyframes spin {
-  from {
-    transform: rotate(0deg);
-  }
-  to {
-    transform: rotate(360deg);
-  }
-}
-/* --- --- */
-.errorMessageContainer {
-  display: none;
-  padding: 1em;
-  width: fit-content;
-  margin: 0 auto;
-}
-
-.errorMessage {
-  width: fit-content;
-  border: 2px solid darkred;
-  background-color: rgb(232, 190, 190);
-  color: darkred;
-  padding: 20px;
-  font-size: 1.5em;
-}
-
-/* --- index.html --- */
-/* Main */
-.generation-title {
-  text-align: center;
-}
-
-/* Container for all Pokemon-cards */
-.pokeCardContainer {
+/* --- --- LANDING PAGE --- --- */
+.card-container {
   display: flex;
   flex-direction: column;
-  gap: 1em;
-  padding: 0.5em;
+  gap: 0.75em;
 }
-.pokeCardContainer a {
+.card-container a {
   text-decoration: none;
 }
-.pokeCardContainer a .pokeCard {
+.card-container a .pokeCard {
   display: flex;
   align-items: center;
   width: 100%;
@@ -188,7 +115,7 @@ html {
   gap: 1em;
   background-color: var(--card-color4);
 }
-.pokeCardContainer a .pokeCard .pokeCard-background {
+.card-container a .pokeCard .pokeCard-background {
   position: absolute;
   z-index: 0;
   background: url(/src/images/pokeball-background.png) no-repeat;
@@ -198,28 +125,28 @@ html {
   height: 4em;
   opacity: 0.2;
 }
-.pokeCardContainer a .pokeCard .pokeCard-image {
+.card-container a .pokeCard .pokeCard-image {
   position: relative;
   z-index: 1;
   width: 4em;
   height: 4em;
   margin: auto 0 auto 0.5em;
 }
-.pokeCardContainer a .pokeCard .pokeCard-text {
+.card-container a .pokeCard .pokeCard-text {
   display: flex;
   flex-direction: column;
 }
-.pokeCardContainer a .pokeCard .pokeCard-text span {
+.card-container a .pokeCard .pokeCard-text span {
   color: black;
 }
-.pokeCardContainer a .pokeCard .pokeCard-text .pokeCard-id {
+.card-container a .pokeCard .pokeCard-text .pokeCard-id {
   font-size: 0.875rem;
 }
-.pokeCardContainer a .pokeCard .pokeCard-text .pokeCard-name {
+.card-container a .pokeCard .pokeCard-text .pokeCard-name {
   font-size: 1rem;
   text-transform: uppercase;
 }
-.pokeCardContainer a:hover {
+.card-container a:hover {
   border-radius: 0.5em;
   box-shadow: 0px 0px 3px rgb(30, 30, 30);
   transition: box-shadow 0.1s ease-in-out;
@@ -242,72 +169,77 @@ html {
   }
 }
 @media screen and (min-width: 550px) {
-  .pokeCardContainer {
+  .card-container {
     display: grid;
     grid-template-columns: 1fr 1fr;
-    gap: 0.5em;
   }
-  .pokeCardContainer .pokeCard {
+  .card-container .pokeCard {
     display: flex;
     width: 100%;
     height: 6em;
   }
 }
-@media screen and (min-width: 678px) {
-  .pokeCardContainer {
+@media screen and (min-width: 768px) {
+  .card-container {
     display: flex;
     flex-flow: row wrap;
     justify-content: center;
-    gap: 1em;
     padding: 0;
   }
-  .pokeCardContainer a .pokeCard {
+  .card-container a .pokeCard {
     flex-direction: column;
     text-align: center;
     width: 14em;
     height: 14em;
   }
-  .pokeCardContainer a .pokeCard .pokeCard-background {
+  .card-container a .pokeCard .pokeCard-background {
     margin-left: 0;
     margin-top: 0.5em;
     width: 7.5em;
     height: 7.5em;
   }
-  .pokeCardContainer a .pokeCard .pokeCard-image {
+  .card-container a .pokeCard .pokeCard-image {
     width: 7.5em;
     height: 7.5em;
     margin: 0 auto;
   }
-  .pokeCardContainer a .pokeCard .pokeCard-text {
+  .card-container a .pokeCard .pokeCard-text {
     display: flex;
     flex-direction: column;
     gap: 0.5em;
   }
-  .pokeCardContainer a .pokeCard .pokeCard-text .pokeCard-id {
+  .card-container a .pokeCard .pokeCard-text .pokeCard-id {
     font-size: 1rem;
     margin: 0 auto;
   }
-  .pokeCardContainer a .pokeCard .pokeCard-text .pokeCard-name {
+  .card-container a .pokeCard .pokeCard-text .pokeCard-name {
     font-size: 1.25rem;
     margin: 0 auto;
   }
 }
-/* --- SELECTED POKEMON --- */
-.pokemonContainer {
+.load-more-pokémon-btn {
+  display: none;
+  width: 100%;
+  padding: 1em;
+  margin: 1.5em auto;
+  font-weight: 600;
+  background-color: rgb(251, 202, 60);
+  color: black;
+}
+.load-more-pokémon-btn:hover {
+  background-color: rgb(230, 185, 55);
+}
+@media screen and (min-width: 768px) {
+  .load-more-pokémon-btn {
+    width: 20em;
+  }
+}
+
+/* --- --- SELECTED POKEMON --- --- */
+.pokemon-container {
   background-color: var(--background-color);
   border-radius: 1em;
   margin-top: 7.5em;
-}
-
-.title_selectedPokemon {
-  text-align: center;
-  text-transform: capitalize;
-  font-weight: 900;
-  font-size: 3rem;
-}
-
-.pokemonId_selectedPokemon {
-  font-size: 1.5rem;
 }
 
 .selectedPokemon-image {
@@ -383,7 +315,7 @@ html {
   height: 4em;
 }
 
-@media screen and (min-width: 678px) {
+@media screen and (min-width: 768px) {
   .selectedPokemon-infoContainer {
     display: grid;
     grid-template-columns: 1fr 1fr;
