+++ src/scss/style.scss
@@ -1,108 +1,17 @@
 @import './global-styles.scss';
 @import './color-schemes.scss';
 
-.contentContainer {
-  width: min(80em, 100% - 1rem);
-  margin-inline: auto;
-}
-
-.loadingContainer {
-  display: none;
-  flex-direction: column;
-  justify-content: center;
-  align-items: center;
-  padding: 2em;
-}
-
-.loadingText {
-  font-size: 1.5em;
-  font-weight: bold;
-}
-
-/* --- Loading animation --- */
-#loadingAnimation {
-  display: flex;
-  position: relative;
-  height: 100px;
-  width: 100px;
-  background: linear-gradient(
-    to bottom,
-    rgb(254, 0, 1) 50%,
-    rgb(248, 248, 248) 50%
-  );
-  box-shadow: inset -8px -4px 0 0 rgba(0, 0, 0, 0.2);
-  border-radius: 50%;
-  border: 5px solid rgb(35, 35, 35);
-  animation: spin 1s linear infinite;
-
-  &:before {
-    content: '';
-    position: absolute;
-    height: 5px;
-    width: 100%;
-    background: rgb(35, 35, 35);
-    top: 50%;
-    transform: translatey(-50%);
-  }
-
-  &:after {
-    content: '';
-    position: absolute;
-    height: 40%;
-    width: 40%;
-    border-radius: 50%;
-    background: rgb(248, 248, 248);
-    top: 50%;
-    left: 50%;
-    transform: translate(-50%, -50%);
-    box-shadow:
-      inset 0 0 0 5px rgb(35, 35, 35),
-      inset 0 0 0 10px rgb(248, 248, 248),
-      inset 0 0 0 12px rgb(35, 35, 35);
+.content {
+  h1 {
+    color: var(--text-color);
   }
 }
 
-/* Spin Animation */
-@keyframes spin {
-  from {
-    transform: rotate(0deg);
-  }
-  to {
-    transform: rotate(360deg);
-  }
-}
-/* --- --- */
-
-.errorMessageContainer {
-  display: none;
-  padding: 1em;
-  width: fit-content;
-  margin: 0 auto;
-}
-
-.errorMessage {
-  width: fit-content;
-  border: 2px solid darkred;
-  background-color: rgb(232, 190, 190);
-  color: darkred;
-  padding: 20px;
-  font-size: 1.5em;
-}
-
-/* --- index.html --- */
-
-/* Main */
-
-.generation-title {
-  text-align: center;
-}
-
-/* Container for all Pokemon-cards */
-.pokeCardContainer {
+/* --- --- LANDING PAGE --- --- */
+.card-container {
   display: flex;
   flex-direction: column;
-  gap: 1em;
-  padding: 0.5em;
+  gap: 0.75em;
 
   a {
     text-decoration: none;
@@ -174,10 +83,9 @@
 }
 
 @media screen and (min-width: 550px) {
-  .pokeCardContainer {
+  .card-container {
     display: grid;
     grid-template-columns: 1fr 1fr;
-    gap: 0.5em;
 
     .pokeCard {
       display: flex;
@@ -187,12 +95,11 @@
   }
 }
 
-@media screen and (min-width: 678px) {
-  .pokeCardContainer {
+@media screen and (min-width: 768px) {
+  .card-container {
     display: flex;
     flex-flow: row wrap;
     justify-content: center;
-    gap: 1em;
     padding: 0;
 
     a {
@@ -233,24 +140,30 @@
     }
   }
 }
+.load-more-pokémon-btn {
+  display: none;
+  width: 100%;
+  padding: 1em;
+  margin: 1.5em auto;
+  font-weight: 600;
+  background-color: rgb(251, 202, 60);
+  color: black;
 
-/* --- SELECTED POKEMON --- */
+  &:hover {
+    background-color: rgb(230, 185, 55);
+  }
 
-.pokemonContainer {
+  @media screen and (min-width: 768px) {
+    width: 20em;
+  }
+}
+
+/* --- --- SELECTED POKEMON --- --- */
+
+.pokemon-container {
   background-color: var(--background-color);
   border-radius: 1em;
   margin-top: 7.5em;
-}
-
-.title_selectedPokemon {
-  text-align: center;
-  text-transform: capitalize;
-  font-weight: 900;
-  font-size: 3rem;
-}
-
-.pokemonId_selectedPokemon {
-  font-size: 1.5rem;
 }
 
 .selectedPokemon-image {
@@ -326,7 +239,7 @@
   height: 4em;
 }
 
-@media screen and (min-width: 678px) {
+@media screen and (min-width: 768px) {
   .selectedPokemon-infoContainer {
     display: grid;
     grid-template-columns: 1fr 1fr;
