+++ tsconfig.json
@@ -1,3 +1,3 @@
 {
-  "extends": "astro/tsconfigs/base"
+  "extends": "astro/tsconfigs/base",
 }
